requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchaudio>=2.0.0
numpy>=1.21.0
Pillow>=9.0.0
matplotlib>=3.5.0
seaborn>=0.11.0
scikit-learn>=1.0.0
tqdm>=4.64.0
streamlit>=1.25.0
opencv-python>=4.7.0
pandas>=1.5.0
plotly>=5.15.0
streamlit-option-menu>=0.3.6